Fix datatype mismatch and add more info on ATSAMx USART registers

+++ src/atsamd/gpio.h
@@ -37,7 +37,7 @@ uint16_t gpio_adc_read(struct gpio_adc g);
 void gpio_adc_cancel_sample(struct gpio_adc g);
 
 struct sercom_buffer {
-  uint32_t *data;
+  uint8_t *data;
   uint32_t maxlen;
   uint32_t start;
   uint32_t len;
@@ -56,8 +56,8 @@ struct usart_config {
 struct usart_config usart_setup(uint32_t bus, uint8_t mode, uint32_t rate, struct usart_buffer* buffer);
 void usart_write_unbuffered(struct usart_config config, uint8_t write_len, uint8_t *write);
 void usart_read_unbuffered(struct usart_config config, uint8_t read_len, uint8_t *read);
-void usart_set_tx_buffer(struct usart_config config, uint32_t* buffer_data, uint8_t buffer_len);
-void usart_set_rx_buffer(struct usart_config config, uint32_t* buffer_data, uint8_t buffer_len);
+void usart_set_tx_buffer(struct usart_config config, uint8_t* buffer_data, uint8_t buffer_len);
+void usart_set_rx_buffer(struct usart_config config, uint8_t* buffer_data, uint8_t buffer_len);
 void usart_write(struct usart_config config, uint8_t write_len, uint8_t *write);
 void usart_read(struct usart_config config, uint8_t read_len, uint8_t *read);
